requirements.txt
argcomplete==3.1.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.28.5
botocore==1.31.5
certifi==2023.5.7
cfn-flip==1.3.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Django==4.2.3
django-phonenumber-field==7.1.0
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
phonenumbers==8.13.16
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.8
Werkzeug==2.3.6
zappa==0.57.0
